+++ CHANGELOG.md
@@ -106,6 +106,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - dird: `list jobs`: add `level` keyword and accept a list of job levels [PR #1548]
 - Allow to use the third-party libraries of the OS instead of the bundled ones [PR #1441]
 - packaging: debian fix dependencies [PR #1573]
+- stored: fix support for non-tape block-addressed devices [PR #1554]
 
 ### Removed
 - remove no longer used pkglists [PR #1335]
@@ -270,6 +271,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 [PR #1548]: https://github.com/bareos/bareos/pull/1548
 [PR #1549]: https://github.com/bareos/bareos/pull/1549
 [PR #1550]: https://github.com/bareos/bareos/pull/1550
+[PR #1554]: https://github.com/bareos/bareos/pull/1554
 [PR #1556]: https://github.com/bareos/bareos/pull/1556
 [PR #1563]: https://github.com/bareos/bareos/pull/1563
 [PR #1565]: https://github.com/bareos/bareos/pull/1565